requirements.txt
# Whale Tracker - Production Dependencies

# Core
asyncio
aiohttp>=3.9.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Blockchain
web3>=6.15.0

# Data & Math
pandas>=2.1.0
numpy>=1.24.0

# Scheduling
APScheduler>=3.10.0

# Validation
pydantic>=2.5.0

# Notifications
python-telegram-bot==20.7  # Для Telegram API

# Utilities
python-dotenv>=1.0.0
